TLDR-backend/application/requirements.txt
absl-py==1.0.0
astroid==2.11.5
astunparse==1.6.3
bert-extractive-summarizer==0.10.1
blis==0.7.7
cachetools==5.1.0
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cymem==2.0.6
dill==0.3.5.1
docopt==0.6.2
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
huggingface-hub==0.6.0
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
lazy-object-proxy==1.7.1
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.7.0
murmurhash==1.0.7
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pathy==0.6.1
pipreqs==0.4.11
platformdirs==2.5.2
preshed==3.0.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pylint==2.13.9
pyparsing==3.0.9
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.53
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
thinc==8.0.16
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.0
